fix some problems with overlay selection dialog
sometimes long overlay names could not be read (as happened on XGA
resolution when presenting ilastik with a beamer). Solution is to use a
QSplitter to split left and right parts.

Remove QScrollArea, wasn't used anywhere.

Remove hardcoded size settings.

Use QSizePolicy.Expanding for the preview, so it get's as large as
possible.